Move method event extension to core change classification project
